+++ R/DOSE.R
@@ -1,8 +1,0 @@
-
-#' @importFrom enrichit geneID
-#' @export
-enrichit::geneID
-
-#' @importFrom enrichit geneInCategory
-#' @export
-enrichit::geneInCategory