+++ R/ReactomePA-package.R
@@ -1,6 +1,9 @@
 #' @keywords internal
 "_PACKAGE"
 
+# Declare ggplot2 aesthetic variables used in viewPathway()
+utils::globalVariables(c("color", "size", "name"))
+
 
 #' Datasets
 #' sample contains a sample of gene IDs.